Move `\SMW\PropertyRestrictionExaminer` to `\SMW\Property\RestrictionExaminer`, refs 3801
