+++ src/Aliases.php
@@ -11,6 +11,7 @@ class_alias( \SMW\Query\ResultPrinters\DsvResultPrinter::class, 'SMWDSVResultPri
 class_alias( \SMW\Query\ResultPrinters\AggregatablePrinter::class, 'SMWAggregatablePrinter' );
 class_alias( \SMW\Property\Annotator::class, 'SMW\PropertyAnnotator' );
 class_alias( \SMW\Property\SpecificationLookup::class, 'SMW\PropertySpecificationLookup' );
+class_alias( \SMW\Property\RestrictionExaminer::class, 'SMW\PropertyRestrictionExaminer' );
 
 // 3.0
 class_alias( \SMW\MediaWiki\Deferred\CallableUpdate::class, 'SMW\DeferredCallableUpdate' );
